ai-legal-assistant/backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database - PostgreSQL with Supabase
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
supabase==2.19.0

# LLM - Gemini
google-generativeai>=0.5.0

# AI and ML
openai==1.3.7
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2
huggingface-hub==0.16.4
pinecone-client==2.2.4

# Document processing
pdfplumber==0.10.3
PyMuPDF>=1.24.2
python-docx==1.1.0
pytesseract==0.3.10
Pillow==10.1.0
pymupdf4llm==0.0.5


# HTTP and API
httpx>=0.26,<0.29
requests==2.31.0
aiofiles==23.2.1

# Data processing
pandas==2.1.3
numpy==1.25.2
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Email
sendgrid==6.10.0
emails==0.6.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
click==8.1.7
rich==13.7.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
